pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "yukkimusic"
version = "1.5.0"
description = "This is YukkiMusic an updated version of YukkiMusicBot for streaming in Telegram voice chats."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
license-files = ["LICENSE"]
authors = [
  { name = "TheTeamVivek" }
]
keywords = ["telegram", "music", "bot", "pyrogram", "pytgcalls"]
classifiers = [
  "Environment :: Console",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Utilities",
]

dependencies = [
    "async-lru==2.0.5",
    "aiofiles==24.1.0",
    "aiohttp==3.12.13",
    "beautifulsoup4==4.13.4",
    "gitpython==3.1.44",
    "heroku3==5.2.1",
    "Kurigram==2.2.5",
    "lyricsgenius==3.6.4",
    "ntgcalls==2.0.4",
    "pillow==11.2.1",
    "pykeyboard2==0.1.6",
    "pymongo==4.13.2",
    "psutil==7.0.0",
    "py-tgcalls==2.2.3",
    "python-dotenv==1.1.0",
    "speedtest-cli==2.1.3",
    "spotipy==2.25.1",
    "TgCrypto==1.2.5",
    "uvloop==0.21.0",
    "Flask==3.1.1",
    "gunicorn==23.0.0",
    "yt-dlp==2025.6.9",
    "py-yt-search==0.3",
    "PyYAML==6.0.2"
]

[tool.setuptools.packages.find]
include = ["yukkimusic", "config", "strings"]

[project.urls]
Repository = "https://github.com/TheTeamVivek/YukkiMusic"
Issues = "https://github.com/TheTeamVivek/YukkiMusic/issues"
Changelog = "https://github.com/TheTeamVivek/YukkiMusic/blob/master/CHANGELOG.md"

[project.scripts]
yukkimusic = "yukkimusic.__main__:main"
